requirements.txt
robotpy==2026.2.1
certifi
robotpy-rev
ntcore
wpilib
phoenix6
robotpy-pathplannerlib
robotpy-commands-v2